Move tests from converter and FakeC8yProxy to tests
